# src/Hexalith.Timesheets.Contracts/openapi/timesheets-capture-contracts.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hexalith Timesheets Capture Contracts",
    "version": "1.0",
    "summary": "Stable contract description for Timesheets time capture, Timesheet Period submission, Activity Type catalog, evidence reads, and metadata descriptors."
  },
...
  "components": {
    "schemas": {
...
      "ConfirmTimeThroughMagicLink": {
        "type": "object",
        "description": "Confirm scoped time after server-side capability validation. The payload carries no authority or attribution fields; the server derives all attribution from the validated capability.",
        "properties": {},
        "additionalProperties": false
      },
      "AdjustTimeThroughMagicLink": {
        "type": "object",
        "description": "Adjust policy-allowed proposed-entry fields after server-side capability validation. Caller payload is restricted to editable proposed-entry values.",
        "required": [ "serviceDate", "durationMinutes", "activityTypeId", "billableState" ],
        "properties": {
          "serviceDate": { "type": "string", "format": "date" },
          "durationMinutes": { "type": "integer", "format": "int32", "minimum": 1 },
          "activityTypeId": { "$ref": "#/components/schemas/ActivityTypeId" },
          "billableState": { "$ref": "#/components/schemas/BillableState" },
          "comment": { "$ref": "#/components/schemas/TimeEntryComment" }
        },
        "additionalProperties": false
      },
...
      "MagicLinkAdjustmentDisplayResponse": {
        "type": "object",
        "description": "Safe external adjustment display details returned only after capability validation. Read-only and editable fields are explicit.",
        "required": [ "proposedDate", "durationMinutes", "durationUnit", "activityTypeId", "activityTypeLabel", "billableState", "targetContext", "editableFields", "readOnlyFields" ],
        "properties": {
          "proposedDate": { "type": "string", "format": "date" },
          "durationMinutes": { "type": "integer", "format": "int32", "minimum": 1 },
          "durationUnit": { "type": "string" },
          "activityTypeId": { "$ref": "#/components/schemas/ActivityTypeId" },
          "activityTypeLabel": { "type": "string" },
          "billableState": { "$ref": "#/components/schemas/BillableState" },
          "targetContext": { "type": "string" },
          "editableFields": { "type": "array", "items": { "type": "string" } },
          "readOnlyFields": { "type": "array", "items": { "type": "string" } },
          "comment": { "type": "string", "nullable": true }
        },
        "additionalProperties": false
      },
      "MagicLinkConfirmationDisplayResponse": {
        "type": "object",
        "description": "Safe external confirmation display details returned only after capability validation. Excludes authority, approval, and sibling-owned state.",
        "required": [ "proposedDate", "durationMinutes", "durationUnit", "activityTypeId", "activityTypeLabel", "billableState", "targetContext" ],
        "properties": {
          "proposedDate": { "type": "string", "format": "date" },
          "durationMinutes": { "type": "integer", "format": "int32", "minimum": 1 },
          "durationUnit": { "type": "string" },
          "activityTypeId": { "$ref": "#/components/schemas/ActivityTypeId" },
          "activityTypeLabel": { "type": "string" },
          "billableState": { "$ref": "#/components/schemas/BillableState" },
          "targetContext": { "type": "string" },
          "comment": { "type": "string", "nullable": true }
        },
        "additionalProperties": false
      },
...
      "TimeEntryComment": {
        "type": "object",
        "required": [ "text", "policy" ],
        "properties": {
          "text": { "type": "string", "minLength": 1, "maxLength": 4096 },
          "policy": { "$ref": "#/components/schemas/TimeEntryCommentPolicy" }
        }
      },
      "TimeEntryCommentPolicy": {
        "type": "object",
        "properties": {
          "internalDisplay": { "$ref": "#/components/schemas/TimesheetsCommentPolicyDecision" },
          "externalConfirmationDisplay": { "$ref": "#/components/schemas/TimesheetsCommentPolicyDecision" },
          "projectionInclusion": { "$ref": "#/components/schemas/TimesheetsCommentPolicyDecision" },
          "exportInclusion": { "$ref": "#/components/schemas/TimesheetsCommentPolicyDecision" },
          "supportDiagnostics": { "$ref": "#/components/schemas/TimesheetsCommentPolicyDecision" },
          "redactionRequirement": { "$ref": "#/components/schemas/TimesheetsCommentRedactionRequirement" },
          "retentionCategory": { "$ref": "#/components/schemas/TimesheetsEvidenceRetentionCategory" }
        }
      },
...
      "ActivityTypeId": { "type": "string", "minLength": 1 },
...
      "BillableState": {
        "type": "string",
        "enum": [ "Unknown", "NonBillable", "Billable" ]
      },
...
      "TimesheetsEvidenceRetentionCategory": {
        "type": "string",
        "enum": [ "Unknown", "TimeEntryEvidence", "CommentText", "ExportRecord", "MagicLinkConfirmationAuditMetadata" ]
      },
...
      "TimesheetsCommentPolicyDecision": {
        "type": "string",
        "enum": [ "Unknown", "Allowed", "Excluded", "Redacted", "PolicyRequired" ]
      },
      "TimesheetsCommentRedactionRequirement": {
        "type": "string",
        "enum": [ "Unknown", "NotRequired", "RequiredBeforeExternalDisclosure", "RequiredBeforeExport" ]
      }
    }
  },
  "paths": {
    "/api/timesheets/magic-links/confirm": {
      "get": {
        "summary": "Display scoped magic-link confirmation details",
        "parameters": [
          { "name": "t", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": {
            "description": "Safe confirmation details",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MagicLinkConfirmationDisplayResponse" } }
            }
          },
          "403": { "description": "Opaque invalid-link response" }
        }
      }
    },
    "/api/timesheets/magic-links/confirm/submit": {
      "post": {
        "summary": "Confirm scoped proposed time through a magic link",
        "parameters": [
          { "name": "t", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmTimeThroughMagicLink" } }
          }
        },
        "responses": {
          "202": { "description": "Confirmation accepted" },
          "403": { "description": "Opaque invalid-link response" }
        }
      }
    },
    "/api/timesheets/magic-links/adjust": {
      "get": {
        "summary": "Display scoped magic-link adjustment details",
        "parameters": [
          { "name": "t", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "responses": {
          "200": {
            "description": "Safe adjustment details",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MagicLinkAdjustmentDisplayResponse" } }
            }
          },
          "403": { "description": "Opaque invalid-link response" }
        }
      }
    },
    "/api/timesheets/magic-links/adjust/submit": {
      "post": {
        "summary": "Adjust scoped proposed time through a magic link",
        "parameters": [
          { "name": "t", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AdjustTimeThroughMagicLink" } }
          }
        },
        "responses": {
          "202": { "description": "Adjustment accepted" },
          "403": { "description": "Opaque invalid-link response" }
        }
      }
    }
  }
}